+++ app/assets/javascripts/application.js
@@ -66,3 +66,4 @@
 // require ABsetup
 //= require filedrag
 //= require builder/surveys
+//= require admin/exports