+++ app/helpers/admin/exports_helper.rb
@@ -1,0 +1,2 @@
+module Admin::ExportsHelper
+end