+++ app/models/surveys/answer_session.rb
@@ -9,8 +9,8 @@ class AnswerSession < ActiveRecord::Base
   has_many :answers
 
   # Validations
-  validates_presence_of :survey_id, :encounter
-  validates_uniqueness_of :survey_id, scope: [:encounter, :user_id, :child_id, :deleted]
+  validates :survey_id, :encounter, presence: true
+  validates :survey_id, uniqueness: { scope: [:encounter, :user_id, :child_id, :deleted] }
 
   # Model Methods
   def completed?
@@ -22,7 +22,7 @@ class AnswerSession < ActiveRecord::Base
   end
 
   def available_for_user_types?(user_types)
-    self.survey.survey_user_types.where(user_type: user_types).count > 0
+    survey.survey_user_types.where(user_type: user_types).count > 0
   end
 
   def process_answer(question, response)
@@ -47,25 +47,25 @@ class AnswerSession < ActiveRecord::Base
 
   def atomic_first_or_create_answer(question)
     begin
-      self.answers.where(question_id: question.id).first_or_create!
+      answers.where(question_id: question.id).first_or_create!
     rescue ActiveRecord::RecordNotUnique, ActiveRecord::RecordInvalid
       retry
     end
   end
 
   def lock!
-    self.answers.update_all state: 'locked'
-    self.update locked: true, locked_at: Time.zone.now
+    answers.update_all state: 'locked'
+    update locked: true, locked_at: Time.zone.now
   end
 
   def unlock!
-    self.answers.where(state: 'locked').update_all(state: 'complete')
-    self.update locked: false, locked_at: nil
+    answers.where(state: 'locked').update_all(state: 'complete')
+    update locked: false, locked_at: nil
   end
 
   # Returns all answer values for a specific question and answer template
   def answer_values(question, answer_template)
-    AnswerValue.joins(:answer).where(answers: { answer_session_id: self.id, question_id: question.id }).where(answer_template_id: answer_template.id)
+    AnswerValue.joins(:answer).where(answers: { answer_session_id: id, question_id: question.id }).where(answer_template_id: answer_template.id)
   end
 
   def percent_completed
@@ -78,8 +78,6 @@ class AnswerSession < ActiveRecord::Base
 
   def destroy
     update_column :deleted, true
-    answers.each do |a|
-      a.destroy
-    end
+    answers.destroy_all
   end
 end