+++ app/uploaders/photo_uploader.rb
@@ -10,7 +10,7 @@ class PhotoUploader < CarrierWave::Uploader::Base
   # Override the directory where uploaded files will be stored.
   # This is a sensible default for uploaders that are meant to be mounted:
   def store_dir
-    "uploads/#{model.class.to_s.underscore}/#{mounted_as}/#{model.id}"
+    File.join(model.class.to_s.underscore, mounted_as.to_s, model.id.to_s)
   end
 
   # Provide a default URL as a default if there hasn't been a file uploaded:
@@ -22,7 +22,7 @@ class PhotoUploader < CarrierWave::Uploader::Base
   # end
 
   # Process files as they are uploaded:
-  process :resize_to_fit => [200, 300]
+  process resize_to_fit: [200, 300]
 
   #
   # def scale(width, height)
@@ -45,5 +45,4 @@ class PhotoUploader < CarrierWave::Uploader::Base
   def filename
     "photo#{File.extname(original_filename)}" if original_filename
   end
-
 end