+++ config/initializers/carrier_wave.rb
@@ -1,7 +1,7 @@
 CarrierWave.configure do |config|
-  if Rails.env.test?
-    config.root = lambda { File.join( Rails.root, 'test', 'support' ) }
-    config.storage = :file
-    config.enable_processing = false
-  end
+  config.root = if Rails.env.test?
+                  File.join Rails.root, 'test', 'support'
+                else
+                  File.join Rails.root, 'carrierwave'
+                end
 end