+++ db/migrate/20151008164458_create_admin_exports.rb
@@ -1,0 +1,17 @@
+class CreateAdminExports < ActiveRecord::Migration
+  def change
+    create_table :admin_exports do |t|
+      t.string :file
+      t.integer :user_id
+      t.integer :total_steps, null: false, default: 0
+      t.integer :current_step, null: false, default: 0
+      t.string :status, null: false, default: 'started'
+      t.datetime :file_created_at
+      t.text :details
+
+      t.timestamps null: false
+    end
+
+    add_index :admin_exports, :user_id
+  end
+end