+++ test/mailers/previews/user_mailer_preview.rb
@@ -1,6 +1,5 @@
 # Preview all emails at http://localhost:3000/rails/mailers/user_mailer
 class UserMailerPreview < ActionMailer::Preview
-
   def forum_digest
     user = User.first
     UserMailer.forum_digest(user)
@@ -43,4 +42,8 @@ class UserMailerPreview < ActionMailer::Preview
     UserMailer.encounter_digest(owner, 84, { "about-me" => { name: "About Me", encounters: [ { name: "Follow Up", answer_sessions_change: 50 } ] }, "more-about-me" => { name: "More About Me", encounters: [ { name: "Seasonal Follow Up", answer_sessions_change: 12 }, { name: "Second Month", answer_sessions_change: 22 } ] } })
   end
 
+  def export_ready
+    export = Admin::Export.first
+    UserMailer.export_ready(export)
+  end
 end