+++ test/models/admin/export_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class Admin::ExportTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end